INT-3265 - Temporarily disable fetching MS365 documentation
We are currently working on improvements to the documentation
automation to support copying images into the docs repo from
individual integration repos. This is a workaround to fix the
MS365 integration.

The MS365 integration will not be fetched from the
`graph-microsoft-365` project until we uncomment the value
in `integrations.config.yaml`. In the meantime, we can make
relevant changes directly inside of the `docs` repo and the
same changes to the `graph-microsoft-365` repo.